requirements-dev.txt
cachetools==5.3.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
filelock==3.12.0
Flask==2.2.3
google-api-core==2.11.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-cloud-secret-manager==2.10.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.54.0
identify==2.5.23
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
nodeenv==1.7.0
oauthlib==3.2.2
platformdirs==3.5.0
pre-commit==3.2.2
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.15
virtualenv==20.23.0
Werkzeug==2.2.3
